third_party/WebKit/LayoutTests/fast/canvas/canvas-fillStyle-no-quirks-parsing.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function(t) {
    ctx = document.createElement('canvas').getContext('2d');
    ctx.fillStyle = "#008000";
    ctx.fillStyle = "800000";
    assert_equals(ctx.fillStyle, "#008000");
}, 'fillStyle should not accept quirks mode hex colors.');
</script>
